requirements.txt
Django==1.8
decorator==4.0.9
ipdb==0.9.0
ipython==4.1.2
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
python-social-auth==0.2.12
python3-openid==3.0.6
ptyprocess==0.5.1
simplegeneric==0.8.1
traitlets==4.1.0